backend/requirements.txt
# Ujamaa DeFi Platform - Backend Dependencies
# MVP Testnet Release

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database (SQLite for dev, PostgreSQL for production)
sqlalchemy==2.0.25
alembic==1.13.1
python-dotenv==1.0.0

# PostgreSQL adapter (for production)
psycopg2-binary==2.9.9

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Blockchain Integration
web3==6.13.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# CORS support (included in FastAPI but explicit for clarity)
fastapi-cors==0.0.6
